Adding comments in the code, in the Dockerfile, docker-compose and refactoring both queries controllers into a single Druid controller.
