+++ app/routes/createRouter.js
@@ -1,6 +1,9 @@
 const glob = require('glob')
 const Router = require('express').Router
-
+/**
+ * Reads the current folder and creates a router with all the defined endpoints.
+ * @returns The mapped router.
+ */
 module.exports = () => glob
     .sync('**/*.js', { cwd: `${__dirname}/` })
     .map(filename => require(`./${filename}`))
