+++ app/routes/dimensions/index.js
@@ -1,5 +1,8 @@
 const Router = require('express').Router
 const dimensionService = require('../../services/dimensionService')
+/**
+ * Defines an endpoint that returns all the dimensions in the datasource.
+ */
 module.exports = Router().get('/dimensions', async (req, res, next) => {
     try {
         const result = await dimensionService.dimensions()
