+++ app/routes/test/hello.js
@@ -1,5 +1,8 @@
 const Router = require('express').Router()
 
+/**
+ * A testing endpoint.
+ */
 module.exports = Router.get('/hello', async(req, res, next) => {
     res.json({"message" : "Hello."})
 })
