+++ app/services/dimensionService.js
@@ -1,13 +1,17 @@
-const sqlQueryController = require('../controller/sqlQueryController')
+const druidController = require('../controller/druidController')
+/**
+ * Queries for the dimensions using SQL. The result is a JSON string, that is parsed.
+ * @returns The parsed JSON string.
+ */
 const dimensions = async () => {
     const body = {
-        "query" : "select dimensions from sys.segments where datasource = ? group by dimensions",
+        "query" : "SELECT dimensions FROM sys.segments WHERE datasource = ? GROUP BY dimensions",
         "parameters" : [{
             "type" : "VARCHAR",
             "value" : "service"
         }]
     }
-    const { data } = await sqlQueryController.sqlQuery(body)
+    const { data } = await druidController.sqlQuery(body)
     return { "dimensions": JSON.parse(data[0].dimensions)}
 }
 module.exports = {
