+++ app/utils/dataFormatter.js
@@ -1,3 +1,8 @@
+/**
+ * Reads the data to group the events by timestamp.
+ * @param result The result returned by Druid. 
+ * @returns The events grouped by timestamp.
+ */
 const groupByTime = result => {
     const timestampOccurrences = result.map(row => row.timestamp)
     const timestampDistinctOccurrences = [...new Set(timestampOccurrences)]
