+++ config/serverConfig.js
@@ -2,6 +2,10 @@ require('dotenv').config();
 const express = require('express');
 const http = require('http');
 const router = require('../app/routes/createRouter')()
+/**
+ * Express setup. Sets the base endpoint to /api and sets the router.
+ * @returns The prepared Express configuration.
+ */
 module.exports = () => {
     const app = express();
     app.use(express.json())
